prov/efa: move definition of packet types to a separate header file
Each packet type need to define a header struct and some functions,
this patch move these definitions into a separate header file
       rxr_pkt_type.h,

also rename the functions included in the header files to add
a rxr_pkt prefix
