+++ prov/efa/Makefile.include
@@ -57,7 +57,8 @@ _efa_headers = \
 	prov/efa/src/rxr/rxr_cntr.h \
 	prov/efa/src/rxr/rxr_rma.h \
 	prov/efa/src/rxr/rxr_msg.h \
-	prov/efa/src/rxr/rxr_pkt_entry.h
+	prov/efa/src/rxr/rxr_pkt_entry.h \
+	prov/efa/src/rxr/rxr_pkt_type.h
 
 efa_CPPFLAGS += \
 	-I$(top_srcdir)/prov/efa/src/ \
