+++ prov/efa/src/rxr/rxr_rma.h
@@ -39,33 +39,6 @@
 
 #include <rdma/fi_rma.h>
 
-struct rxr_readrsp_hdr {
-	uint8_t type;
-	uint8_t version;
-	uint16_t flags;
-	/* end of rxr_base_hdr */
-	uint8_t pad[4];
-	uint32_t rx_id;
-	uint32_t tx_id;
-	uint64_t seg_size;
-};
-
-struct rxr_readrsp_pkt {
-	struct rxr_readrsp_hdr hdr;
-	char data[];
-};
-
-static inline struct rxr_readrsp_hdr *rxr_get_readrsp_hdr(void *pkt)
-{
-	return (struct rxr_readrsp_hdr *)pkt;
-}
-
-#define RXR_READRSP_HDR_SIZE	(sizeof(struct rxr_readrsp_hdr))
-
-#if defined(static_assert) && defined(__x86_64__)
-static_assert(sizeof(struct rxr_readrsp_hdr) == sizeof(struct rxr_data_hdr), "rxr_readrsp_hdr check");
-#endif
-
 struct rxr_rma_read_info {
 	uint64_t rma_initiator_rx_id;
 	uint64_t window;
@@ -75,10 +48,10 @@ struct rxr_rma_read_info {
 static_assert(sizeof(struct rxr_rma_read_info) == 16, "rxr_rma_read_hdr check");
 #endif
 
-char *rxr_rma_init_rts_hdr(struct rxr_ep *ep,
-			   struct rxr_tx_entry *tx_entry,
-			   struct rxr_pkt_entry *pkt_entry,
-			   char *hdr);
+char *rxr_pkt_init_rts_rma_hdr(struct rxr_ep *ep,
+			       struct rxr_tx_entry *tx_entry,
+			       struct rxr_pkt_entry *pkt_entry,
+			       char *hdr);
 
 int rxr_rma_verified_copy_iov(struct rxr_ep *ep, struct fi_rma_iov *rma,
 			      size_t count, uint32_t flags, struct iovec *iov);
@@ -100,19 +73,19 @@ struct rxr_tx_entry *
 rxr_rma_alloc_readrsp_tx_entry(struct rxr_ep *rxr_ep,
 			       struct rxr_rx_entry *rx_entry);
 
-int rxr_rma_init_readrsp_pkt(struct rxr_ep *ep,
-			     struct rxr_tx_entry *tx_entry,
-			     struct rxr_pkt_entry *pkt_entry);
+int rxr_pkt_init_readrsp(struct rxr_ep *ep,
+			 struct rxr_tx_entry *tx_entry,
+			 struct rxr_pkt_entry *pkt_entry);
 
-void rxr_rma_handle_readrsp_sent(struct rxr_ep *ep,
+void rxr_pkt_handle_readrsp_sent(struct rxr_ep *ep,
 				 struct rxr_pkt_entry *pkt_entry);
 
 /* EOR related functions */
-int rxr_rma_init_eor_pkt(struct rxr_ep *ep,
-			 struct rxr_rx_entry *rx_entry,
-			 struct rxr_pkt_entry *pkt_entry);
+int rxr_pkt_init_eor(struct rxr_ep *ep,
+		     struct rxr_rx_entry *rx_entry,
+		     struct rxr_pkt_entry *pkt_entry);
 
-void rxr_rma_handle_eor_sent(struct rxr_ep *ep,
+void rxr_pkt_handle_eor_sent(struct rxr_ep *ep,
 			     struct rxr_pkt_entry *pkt_entry);
 
 size_t rxr_rma_post_shm_rma(struct rxr_ep *rxr_ep,
